Base class pseudo 3d
